@@ -1,13 +1,8 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000"
+        "dev": "vite",
+        "prod": "vite build"
     },
     "dependencies": {
         "apexcharts": "^3.35.3",
@@ -28,28 +23,21 @@
         "@babel/plugin-proposal-optional-chaining": "^7.17.12",
         "@mdi/font": "^6.7.96",
         "@popperjs/core": "^2.10.2",
-        "browser-sync": "^2.27.10",
-        "browser-sync-webpack-plugin": "^2.3.0",
-        "css-minimizer-webpack-plugin": "^4.0.0",
-        "dotenv-webpack": "^7.1.0",
+        "@vitejs/plugin-vue": "^2.3.3",
         "eslint": "^7.8.1",
         "eslint-plugin-prettier": "^3.1.4",
         "eslint-plugin-you-dont-need-lodash-underscore": "^6.10.0",
-        "laravel-mix": "^6.0.6",
-        "mini-css-extract-plugin": "^2.6.0",
+        "laravel-vite-plugin": "^0.2.4",
         "moment": "^2.27.0",
         "moment-duration-format": "^2.3.2",
         "moment-timezone": "^0.5.31",
-        "postcss": "^8.1.14",
-        "resolve-url-loader": "^5.0.0",
-        "sass": "^1.52.1",
-        "sass-loader": "^12.6.0",
+        "postcss": "^8.4.14",
+        "vite": "^2.9.13",
         "vue-cli-plugin-vuetify": "~2.5.0",
-        "vue-loader": "^16.8.3",
-        "webpack-plugin-vuetify": "^2.0.0-alpha.0"
+        "vue-loader": "^16.8.3"
     },
     "engines": {
-        "node": ">= 16.5.0",
-        "npm": ">= 8.5.5"
+        "node": ">= 18.4.0",
+        "npm": ">= 8.12.1"
     }
 }
