@@ -11,9 +11,7 @@
 
 <div id="app"></div>
 
-<script src="{{ mix('js/manifest.js') }}"></script>
-<script src="{{ mix('js/vendor.js') }}"></script>
-<script src="{{ mix('js/admin.js') }}"></script>
+@vite('resources/js/admin.js')
 
 </body>
 </html>
